🐛 화면 버그 수정
