--- src/components/reward/main.tsx
@@ -1,3 +1,0 @@
-export default function main() {
-  return <div>main</div>;
-}
